+++ Controls-demo/treeGridNew/NodeFooter/Configuration/CustomFooter/CustomFooter.wml
@@ -1,4 +1,5 @@
 <div class="controlsDemo__wrapper controlsDemo__treeGrid__configuration controlsDemo_Configuration__grid-four controlsDemo__maxWidth500">
+    <Controls.buttons:Button attr:data-qa="reset-button" caption="Reset callbacks" on:click="_resetCallbacks()"/>
     <Controls.treeGrid:View
             keyProperty="key"
             name="tree4"
@@ -8,6 +9,9 @@
             nodeProperty="type"
             navigation="{{_navigation}}"
             rowSeparatorSize="s"
-            columnSeparatorSize="s">
+            columnSeparatorSize="s"
+            itemActions="{{_itemActions}}"
+            nodeFooterVisibilityCallback="{{_nodeFooterVisibilityCallback}}"
+            itemActionVisibilityCallback="{{_itemActionVisibilityCallback}}">
     </Controls.treeGrid:View>
 </div>
