+++ Controls-demo/treeGridNew/NodeFooter/NodeFooter.wml
@@ -25,13 +25,4 @@
         </Router.router:Reference>
         <Controls-demo.treeGridNew.NodeFooter.MoreButton.Index/>
     </div>
-
-    <div class="controlsDemo__cell">
-        <Router.router:Reference attr:class="controls-text-label"
-                                 state="/Controls-demo/app/:app"
-                                 app="{{ 'Controls-demo/treeGridNew/NodeFooter/NodeFooterVisibilityCallback/Index' }}">
-            <a href="{{content.href}}">Функция обратного вызова для настройки видимости подвала узла</a>
-        </Router.router:Reference>
-        <Controls-demo.treeGridNew.NodeFooter.NodeFooterVisibilityCallback.Index/>
-    </div>
 </div>
